requirements.txt
#pip3 install -r requirements.txt
python-dotenv
streamlit
ffmpeg-python
pydub
huggingface_hub
pyannote.audio
chardet
#git+https://github.com/openai/whisper.
#conda install pytorch torchvision torchaudio pytorch-cuda=11.7 -c pytorch -c nvidia
#torch==1.11.0 
#torchvision==0.12.0 
#torchaudio==0.11.0
#numpy==1.24

#git+https://github.com/m-bain/whisperx.git
#conda install pytorch==2.0.0 torchaudio==2.0.0 pytorch-cuda=11.8 -c pytorch -c nvidia
